FrontendHandler: get be path from fe entry instead of creating manually
As we can have different device name for fe and be, we should read be
path from fe entry instead of creating it. It allows to remove be dom id
parameter form FrontendHandlerBase. As interface is changed, after this
commit, all backends have to be updated according to new interface.

Signed-off-by: Oleksandr Grytsov <oleksandr_grytsov@epam.com>