XenGnttabBuffer: handle data offset for dmabufs
Handle additional data offset parameter while exporting and
importing dmabufs: dmabuf is backed by a scatter-gather table
and has offset parameter which tells where the actual data starts:
  - when dmabuf is created (exported) from grant references then
    data_ofs is used to set the offset field in the scatter list
    of the new dma-buf
  - when dma-buf is imported and then grant references provided
    to the user space then data_ofs is used to report that offset
    to user-space

Note! There is a dependency on libgnttab if it supports importing/
exporting dmabufs with data offset, so conditional compilation is
employed: new definition GNTTAB_HAS_DMABUF_OFFSET is set if the
library supports the offset. If it doesn't and a non-zero offset is
requested/provided, then a run-time exception is raised.

Signed-off-by: Oleksandr Andrushchenko <oleksandr_andrushchenko@epam.com>
Reviewed-by: Oleksandr Grytsov <oleksandr_grytsov@epam.com>
Reviewed-by: Volodymyr Babchuk <volodymyr_babchuk@epam.com>